# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.CategoryResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  dto.Meta:
    properties:
      currentPage:
        type: integer
      limit:
        type: integer
      totalPages:
        type: integer
      totalRecords:
        type: integer
    type: object
  dto.PaginationResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/dto.Meta'
    type: object
...
  dto.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Backend API for SmartFarming fullstack application using Gin, GORM,
    PostgreSQL, MongoDB, and Redis.
  title: SmartFarming API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Lists categories with pagination and search parameters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Search by name or description
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Categories with search and pagination
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Creates a new article category (Admin/Operator only)
      parameters:
      - description: Create Category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a Category
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a category by its UUID (Admin/Operator only)
      parameters:
      - description: Category UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Category
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Fetches details of a category by its UUID
      parameters:
      - description: Category UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Get Category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Updates category details by its UUID (Admin/Operator only)
      parameters:
      - description: Category UUID
        in: path
        name: id
        required: true
        type: string
      - description: Update Category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Category
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"